seperate out the search box of listeners to another component
